Breaking Change: Reconfigure struct stat.
Since there are a couple other breaking changes coming and a version bump
just happened, it seemed like a good time to make a core structure change.

This change rearranges struct stat to alias nicely with FILE_PROPERTIES.
This cuts down on copying and translating of structs. It also adds birth
time, flags, a generation number, and nanoseconds to the stat structure.
These are currently unused but will hopefully be used by future file
systems.

Additionally, this change converts the SET_FILE_INFORMATION.FileProperties
member into a pointer. This also cuts down on copying for get/set file
properties.

FILE_PROPERTIES have also changed a bit as well. FileSize is now Size and
is no longer an INT64_SYNC, since the file object lock properly protects
against all access to the file properties. Tearing is no longer a concern.

Users will need to completely rebuild third-party (including tools) after
this change, as older applications won't work with a misaligned struct
stat.

+++ apps/libc/include/sys/types.h
@@ -53,7 +53,7 @@ typedef long long blkcnt_t;
 // Define a type used for block sizes.
 //
 
-typedef unsigned long blksize_t;
+typedef unsigned long long blksize_t;
 
 //
 // Define a type for file system block counts.
